xspectrum1d stuff
