+++ linetools/spectra/__init__.py
@@ -1,2 +1,2 @@
-import io
-import utils
+#import io
+#import utils
