+++ linetools/spectra/io.py
@@ -17,7 +17,6 @@ from astropy.io.fits.fitsrec import FITS_rec
 from astropy.io.fits.hdu.table import BinTableHDU
 
 #from xastropy.xutils import xdebug as xdb
-from linetools.spectra.utils import XSpectrum1D
 
 #### ###############################
 #  Generate Spectrum1D from FITS file
@@ -42,7 +41,7 @@ def readspec(specfil, inflg=None, efil=None, verbose=False, flux_tags=None,
       Default: flux_tags = ['SPEC', 'FLUX','FLAM','FX', 'FLUXSTIS', 'FLUX_OPT', 'fl']
     sig_tags: list of strings, optional
       Tags for error in Binary FITS table
-      Default: flux_tags = ['SPEC', 'FLUX','FLAM','FX', 'FLUXSTIS', 'FLUX_OPT', 'fl']
+      Default: sig_tags = ['ERROR','ERR','SIGMA_FLUX','FLAM_SIG', 'SIGMA_UP', 'ERRSTIS', 'FLUXERR', 'er']
     multi_ivar: bool, optional 
       If True, assume BOSS format of  flux, ivar, log10(wave) in multi-extension FITS
 
@@ -51,6 +50,7 @@ def readspec(specfil, inflg=None, efil=None, verbose=False, flux_tags=None,
     A Spectrum1D class (or XSpectrum1D which is an overloaded variant)   
     '''
     from specutils.io import read_fits as spec_read_fits
+    from linetools.spectra.xspectrum1d import XSpectrum1D
 
     # Initialize
     dat = None
@@ -58,7 +58,7 @@ def readspec(specfil, inflg=None, efil=None, verbose=False, flux_tags=None,
         inflg = 0
 
     # Check specfil type
-    if isinstance(specfil,Table):
+    if isinstance(specfil,Table):  # MAYBE SHOULD USE SPECUTILS FROM_TABLE
         datfil = 'None'
         # Dummy hdulist
         hdulist = [fits.PrimaryHDU(), specfil]
